Refactoring quality review 1
