+++ app/src/main/java/com/cocoon/movies/moviereview/activites/MainActivity.java
@@ -1,47 +1,42 @@
-package com.example.muhammad.moviereview;
+package com.cocoon.movies.moviereview.activites;
 
 import android.app.LoaderManager;
+import android.app.LoaderManager.LoaderCallbacks;
 import android.content.Context;
 import android.content.Intent;
 import android.content.Loader;
-import android.content.SharedPreferences;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.net.Uri;
-import android.preference.PreferenceManager;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.ListView;
-import android.app.LoaderManager.LoaderCallbacks;
 import android.widget.ProgressBar;
 import android.widget.SearchView;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import com.cocoon.movies.moviereview.R;
+import com.cocoon.movies.moviereview.adapters.MovieAdapter;
+import com.cocoon.movies.moviereview.api.MovieLoader;
+import com.cocoon.movies.moviereview.dto.Movie;
+import com.cocoon.movies.moviereview.utils.QueryUtils;
+import com.google.common.base.Preconditions;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class MainActivity extends AppCompatActivity implements LoaderCallbacks<List<Movie>> {
-    public static final String LOG_TAG = MainActivity.class.getName();
-
-    private final String NYT_MOVIE_REVIEWS_LINK = "https://api.nytimes.com/svc/movies/v2/reviews/search.json";
-    private final String API_KEY = "65ec98d0a3ea4a72ae6a068e1eaa8f3f";
-    private  String QUERY_MOVIE = "";
+    private static final String LOG_TAG = MainActivity.class.getName();
     private static final int MOVIE_LOADER_ID = 1;
-
     private MovieAdapter mAdapter;
-
+    private String movieQuery = "";
     private TextView mEmptyStateTextView;
-
     private ProgressBar mProgressBar;
-
     private SearchView mSearchView;
 
 
@@ -49,7 +44,7 @@ public class MainActivity extends AppCompatActivity implements LoaderCallbacks<L
     public Loader<List<Movie>> onCreateLoader(int i, Bundle bundle) {
         Log.d(LOG_TAG, "onCreateLoader");
 
-        String queryURL = formatQueryLink(QUERY_MOVIE);
+        String queryURL = QueryUtils.formatQueryLink(this, movieQuery);
         return new MovieLoader(this, queryURL);
 
     }
@@ -63,27 +58,23 @@ public class MainActivity extends AppCompatActivity implements LoaderCallbacks<L
     @Override
     public void onLoadFinished(Loader<List<Movie>> loader, List<Movie> movies) {
         mAdapter.clear();
-        try {
-            if (movies != null || !movies.isEmpty()) {
-                mAdapter.addAll(movies);
-            }
-        }catch(NullPointerException e){
-            Log.e(LOG_TAG, "Null pointer Exception");
+        if (movies != null) {
+            mAdapter.addAll(movies);
         }
         Log.d(LOG_TAG, "OnLoadFinished");
 
     }
-//option menu
+
+    @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.menu, menu);
         return true;
     }
-//click event on option menu
-//to direct to bookmarks activity
+
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        Intent intent=new Intent(this,bookmarks.class);
+        Intent intent = new Intent(this, Bookmarks.class);
         startActivity(intent);
         return super.onOptionsItemSelected(item);
     }
@@ -97,13 +88,13 @@ public class MainActivity extends AppCompatActivity implements LoaderCallbacks<L
         ListView MovieListView = (ListView) findViewById(R.id.list);
 
         // Create a new {@link ArrayAdapter} of Movies.
-         mAdapter = new MovieAdapter(this, new ArrayList<Movie>());
+        mAdapter = new MovieAdapter(this, new ArrayList<Movie>());
 
         // Set the adapter on the {@link ListView}
         // so the list can be populated in the user interface
         MovieListView.setAdapter(mAdapter);
 
-        mSearchView = findViewById(R.id.sMovie_bar);
+        SearchView mSearchView = findViewById(R.id.sMovie_bar);
 
         mSearchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
             @Override
@@ -113,43 +104,31 @@ public class MainActivity extends AppCompatActivity implements LoaderCallbacks<L
 
             @Override
             public boolean onQueryTextChange(String s) {
-                if(!TextUtils.isEmpty(s)){
-                    QUERY_MOVIE = s;
-                    SearchMovie();
+                if (!TextUtils.isEmpty(s)) {
+                    movieQuery = s;
+                    loadMovies();
                 }
                 return false;
             }
         });
 
-
         // testing for network connectivity
         ConnectivityManager cm = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+        Preconditions.checkNotNull(cm);
         NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
         boolean isConnected = activeNetwork != null && activeNetwork.isConnectedOrConnecting();
 
-        if(isConnected) {
+        if (isConnected) {
             LoaderManager loaderManager = getLoaderManager();
-
             loaderManager.initLoader(MOVIE_LOADER_ID, null, this);
-        }else{
+        } else {
             mProgressBar.setVisibility(View.GONE);
-            mEmptyStateTextView.setText("Please, check your connection!");
+            mEmptyStateTextView.setText(R.string.checkYourConnection);
         }
     }
 
-    private String formatQueryLink(String query){
-
-        Uri baseUri = Uri.parse(NYT_MOVIE_REVIEWS_LINK);
-        Uri.Builder uriBuilder = baseUri.buildUpon();
-
-        uriBuilder.appendQueryParameter("api-key", API_KEY);
-        uriBuilder.appendQueryParameter("query", query);
-        return uriBuilder.toString();
-    }
-
-    private void SearchMovie(){
+    private void loadMovies() {
         LoaderManager loaderManager = getLoaderManager();
-
         loaderManager.restartLoader(MOVIE_LOADER_ID, null, this);
     }
 }
