+++ app/src/main/java/com/cocoon/movies/moviereview/dao/Bookmark.java
@@ -1,0 +1,8 @@
+package com.cocoon.movies.moviereview.dao;
+
+import com.cocoon.movies.moviereview.dto.Movie;
+
+public class Bookmark {
+    Movie movie;
+    String url;
+}
