+++ app/src/main/java/com/cocoon/movies/moviereview/tasks/DownloadImageTask.java
@@ -1,0 +1,38 @@
+package com.cocoon.movies.moviereview.tasks;
+
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.os.AsyncTask;
+import android.os.Build;
+import android.support.annotation.RequiresApi;
+import android.util.Log;
+import android.widget.ImageView;
+
+import java.io.InputStream;
+
+public class DownloadImageTask extends AsyncTask<String, Void, Bitmap> {
+    ImageView bmImage;
+
+    public DownloadImageTask(ImageView bmImage) {
+        this.bmImage = bmImage;
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
+    protected Bitmap doInBackground(String... urls) {
+        String urldisplay = urls[0];
+        Bitmap mIcon11 = null;
+        try (InputStream in = new java.net.URL(urldisplay).openStream()) {
+            mIcon11 = BitmapFactory.decodeStream(in);
+
+        } catch (Exception e) {
+            Log.e("Error", e.getMessage());
+            e.printStackTrace();
+        }
+        return mIcon11;
+    }
+
+    protected void onPostExecute(Bitmap result) {
+        bmImage.setImageBitmap(result);
+    }
+}
+
