+++ app/src/main/res/layout/activity_bookmarks.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.muhammad.moviereview.bookmarks">
+    tools:context="com.cocoon.movies.moviereview.bookmarks">
 
 </android.support.constraint.ConstraintLayout>
