+++ app/src/main/res/menu/menu.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.muhammad.moviereview.MainActivity">
+    tools:context="com.cocoon.movies.moviereview.activites.MainActivity">
     <item
         android:id="@+id/bMark"
         android:title="Bookmarks"
