+++ app/src/main/res/raw/config.properties
@@ -1,0 +1,2 @@
+api_url=https://api.nytimes.com/svc/movies/v2/reviews/search.json
+api_key=65ec98d0a3ea4a72ae6a068e1eaa8f3f
