+++ app/src/test/java/com/cocoon/movies/moviereview/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.muhammad.moviereview;
+package com.cocoon.movies.moviereview;
 
 import org.junit.Test;
